create a slightly better organized source folder
